+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 4.0.0
 - This version is _only_ compatible with V13+ (and therefore system version 5.0.0+)
 - All UI has been moved from AppV1 Svelte to AppV2 Handlebars. This was a large undertaking, so there may be some issues, which I will endeavor to fix quickly
+- Added support for the Arcane Grimoire item (increases Arcane Recovery spell slot level regain by 1)
 
 ## Version 3.3.1
 - Fixed bug where hp recovery was capped to `hp.max` even if there was a `hp.tempmax` value